refactor(app): move table storage labels from present to policy
