+++ src/app/policy/mod.rs
@@ -3,4 +3,5 @@ pub(crate) mod password_masking;
 pub(crate) mod preview_cell_text;
 pub mod sql;
 pub mod sqlite_path;
+pub mod table_storage;
 pub mod write;
